Authentication generation (#74)

+++ waspc/examples/todoApp/ext/actions.js
@@ -1,6 +1,15 @@
 import HttpError from '@wasp/core/HttpError.js'
+import { createNewUser } from '@wasp/core/auth.js'
+
+export const signUp = async (args, context) => {
+  await createNewUser({ email: args.email, password: args.password })
+}
 
 export const createTask = async (task, context) => {
+  if (!context.user) {
+    throw new HttpError(403)
+  }
+
   const Task = context.entities.Task
   /*
   if (Math.random() < 0.5) {
@@ -16,6 +25,10 @@ export const createTask = async (task, context) => {
 }
 
 export const updateTaskIsDone = async ({ taskId, newIsDoneVal }, context) => {
+  if (!context.user) {
+    throw new HttpError(403)
+  }
+
   const Task = context.entities.Task
   return Task.update({
     where: { id: taskId },
@@ -24,6 +37,10 @@ export const updateTaskIsDone = async ({ taskId, newIsDoneVal }, context) => {
 }
 
 export const deleteCompletedTasks = async (args, context) => {
+  if (!context.user) {
+    throw new HttpError(403)
+  }
+
   const Task = context.entities.Task
   await Task.deleteMany({
     where: { isDone: true }
@@ -31,6 +48,10 @@ export const deleteCompletedTasks = async (args, context) => {
 }
 
 export const toggleAllTasks = async (args, context) => {
+  if (!context.user) {
+    throw new HttpError(403)
+  }
+
   const Task = context.entities.Task
   const notDoneTasksCount = await Task.count({ where: { isDone: false } })
